run_benchmark/requirements.txt
# WebVoyager Evaluation Requirements

# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0

# Browser automation (install from local clone or PyPI when available)
# browser-use>=0.1.0

# Optional: Additional analysis tools
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
